src/requirements.txt
azure-identity>=1.12.0
azure-mgmt-compute>=30.0.0
azure-mgmt-resource>=23.0.1
azure-mgmt-network>=25.1.0
python-dotenv>=1.0.0